requirements.txt
pandas>=2.0
geopandas>=0.14
shapely>=2.0
requests>=2.31
polyline>=2.0
folium>=0.15
pyyaml>=6.0
tqdm>=4.65
click>=8.1
